MiniBuilder#build(name, ctx), tdb:... cmd args
MiniBuilder#build takes a MiniContext instead of a 'stack trace'.

This way build rules can pass the same context they were given
back to the builder to explicitly build other things.

This changes the API, so should be a 0.4.0 thing.

Command arguments can be of the form 'tdb:...',
which are handled specially.
If you want to pass an argument without special interpretation,
say "tdb:literal:...".

Maybe these special argument should be some other object
instead of strings.  Maybe for 0.5.0.